Use prefix
Peers only store an unlimited amount of other peers with whom they share
a prefix (current hardcoded prefix length of 2). These are the sync
peers, the ones that will later on exchange updates about records
they're responsible for. For now, these records are references to each
other (and there are no updates).

In addition, every peer stores, for every subprefix of its prefix that
diverges in the least significant bit, a limited number of query peers.
These are the ones to which queries will be issued about records (i.e.
currently references) concerning peers that are not sync peers.

In order to obtain that information, query peers must answer queries.
This is not yet implemented.