Send messages via events
Instead of calling methods on other peers in order to communicate with
them, peers now schedule events that call those methods. These can be
delayed to simulate a transmission delay (which is currently fixed at
0.1 time steps for all transmissions).

Scheduling a SendQuery event represents sending a query for an ID to
another peer. When triggered, the event calls the recv_query() method of
the recipient. Similarly, SendResponse represents sending a response to
a query.

Since responses are no longer instantaneous, each peer keeps track of
all queries it has sent out in order to match them to the responses.

Peers currently expect all queries to be answered, timeouts have yet to
be implemented.