Use timeouts
When sending a query, peers now start a timeout (currently set to 2 time
units), at the end of which they no longer wait for a response and send
out a query to the next peer with a prefix closer to the desired record.

In case there are no more such peers, the query is considered to have
failed due to a timeout. Similarly, if the last such peer answers that
he doesn't have the desired record, the query is also considered a
failure.

The timeout is realized by starting a new process (within the simulation
framework) which waits for the specified duration and then reports
failure unless it is interrupted (this "defusing" it). This interrupt is
only triggered when a response is received. A handle to the timeout
process is stored in the PendingQuery object pertaining to the query.