Use query groups
Instead of maintaining a list of query peers, each peer now has a set of
query groups, in which it finds query peers.

When looking up a record, peers now have to go through their query
groups to find a peer whose prefix is closer to the target ID than their
own. This is less performant than the dictionary lookup from before, but
can be sped up.

When introduced to a new peer, a peer will first check if knowing that
peer is useful, i.e. if that peer can serve a prefix this peer needs but
this peer knows less than Peer.MIN_DESIRED_QUERY_PEERS serving that
prefix. If so, this peer will try to join one of the groups that peer is
in, unless they all have at least Peer.MAX_DESIRED_GROUP_SIZE members.
In that case, this peer will create a new group and add the other peer
to it.

Joining groups and adding peers is currently done by simply adding
references to peers to sets. There is not yet any mechanism to request
membership in or send invites to a group. This, of course, has to be
added later, but it isn't clear who decides who gets to approve requests
to join a group, and by what authority someone gets to send out invites.