Inform of failure
When a query ultimately fails (due to either timeout or the last known
peer who should have the record not having it), a peer will now also
inform those who queried him for the record of this failure. This was
not the case previously, queries would always time out. (Of course no
queries ever failed, so this never became an issue.)