Factor out decision points
Moved functionality that could be replaced by peers as they adapt their
behavior into separate functions. Right now they contain the same
behavior, later on there will be multiple implementations for each. What
is currently the function will then be a member variable that is set to
the current functionality.

The decision points are:
- A query for self arrives.
- A query for a sync peer arrives.
- A query for some other peer arrives.
- A successful response arrives.
- A failed response (last peer didn't have the record) arrives.
- A failed response arrives, but there are more peers that could have
  the record.
- A query timed out, and there are no more peers to contact.
- A query timed out, but there are peers left to contact that could have
  the record.