Track reputation
A QueryGroup now holds a dictionary mapping all members to a reputation
score. For now, this is a central instance everyone has access to, so
there's no need to send any messages.

Peers apply rewards or penalties upon receiving successful or failed
query responses or upon a query timing out.

Handling of timeouts is still insufficient: When a response doesn't
arrive before the timeout, the peer applies a penalty and sends a query
to the next peer in line. But if the first queried peer's response
arrives before the second one's, a reward will be assigned to the first
peer (and the first peer only). Querying peers must keep track of whom
they sent queries to in order to apply rewards and penalties correctly.

Peers also need to take into account that responses to their queries are
supposed to be delayed if they have low reputation and adjust their
timeouts accordingly.