Check responding peer
When receiving a response, peers now check that they have actually sent
a query to that peer, and when, in order to properly credit the
responder with reputation.

The PendingQuery object holding information about a query now includes a
dictionary keeping track to what peers queries were sent. Upon
successful response to the query, this object is archived if there are
peers to whom this query has been sent but who haven't responded yet. If
they then respond late, they can still be credited reputation. The
archived PendingQuery is removed after a timeout of
COMPLETED_QUERY_RETENTION_TIME.

There is no difference in the reputation credited when a response
arrives after the query has already been answered.