Apply penalties
Peers now delay responses based on the reputation of the one querying.
They also adjust their timeouts to account for the delayed responses
they expect they're getting.

For now, the delay is 10 time units minus the reputation (restricted to
the interval between 0 and 10). With the current setup with 32 peers,
this leads to the maximum delay of 10 being applied for everyone. If the
delay is not capped, it will actually continue to rise for some peers.