Support queries for prefixes
Peers can now query for only a prefix of an ID. The response should then
be a peer whose ID starts with that prefix. This allows peers to learn
about other peers based on their prefix, in order to e.g. find query
groups.

Currently, no particular such peer is returned if there are multiple
candidates, peers will choose whichever one is most convenient. As an
extreme example, peers will return themselves, if their own ID starts
with the requested prefix. This is of course not very useful, as the
purpose of this feature is to learn about new peers. Useful default
behavior would probably be to select one of the matching peers randomly.
It may also be useful to query for more than one peer at a time.

As a consequence, it is now possible for a peer to have multiple queries
with the same prefix going on (some of them being longer, i.e. more
specific than others).

So PendingQuery now stores all the prefixes that were queried for each
peer that needs to be notified when a response arrives, instead of just
a list of the peers.

Also, send_response (and the related SendResponse class) store a set of
all queried prefixes rather than just one ID. The peer who receives the
response then gets one peer along with a set of all the prefixes this
one peer satisfies. Unfortunately, this means upon receiving a response,
the peer must go through this set and find the prefix that is the key
for the pending query object he has stored. Previously, he could just
use the single queried ID parameter as the key, so this has become
somewhat ugly.