Introduce only a sample of peers
Instead of introducing each peer to every other peer, only introduce
each to a small number (hardcoded to 8). Also increase the number of
peers to 64 for now.

Only knowing 8 other peers means lots of queries are impossible right
now because there is no one to send them to. This means peers have to
first establish a network by making partial queries for prefixes to
which they're not connected. It should be interesting to see if the
network is segmented.