Create peer graph
Use networkx to create a directed graph representing which peer has
knowledge of other peers. Peers are passed this graph in the constructor
and maintain it themselves as they are introduced to new peers.

Before the simulation starts, some connectivity information is printed:
The sync groups (i.e. the IDs of peers for each prefix), the query
groups, as well as the number of nodes, diameter and degree histogram of
each of the strongly connected components.

With the current setup there is only one strongly connected component,
and yet some peers fail to resolve some names because they don't know
any peer closer to the target ID. This can happen because a path to the
peer with a target ID in the connectivity graph doesn't necessarily lead
only over peers whose prefix has a closer match than the querying peer.
But those are the only ones queried, since no peer is aware of the
complete network.