Fix joining query groups
In some cases, it was possible that join_group_with() didn't actually
get the peer into a query group with the other peer. This happened if
the other peer had query groups, but they were all full.

Also changed the behavior so that a peer will first try to join one of
the other peers' groups, then try to add the other peer to one of his
own, and only if that fails to create a new one. Previously, adding the
other to an existing group was never attempted.

As a result, fewer query groups are created during the introduction
phase, and the existing ones have more members. The connectivity graph
now also has 2 components with the current configuration, rather than
the 1 from before. Having more groups seems to, at least in this case,
create more connectedness.