Find peers for uncovered subprefixes
After the initial introduction phase, each peer now checks if there are
any subprefixes for which he doesn't know any responsible peers, and
queues up queries for those prefixes.

Since he may not know any responsible peers even to resolve those
prefixes, all peers can be queried, rather than just those that are
closer to the prefix than him (this includes sync peers). This is done
by calling act_query() directly with the new query_all parameter, rather
than calling handle_request() or recv_query().

Once a successful response arrives, the peer calls introduce() on it and
thus can use this peer to resolve queries under the previously uncovered
subprefix. In fact, introduce() is called on every successful response,
so that normal queries supply more connections to a peer. introduce()
doesn't change anything if the peer already knows the peer from the
response, or there are enough query peers for that prefix already, so
it's not a problem calling it on every peer that comes by. Later on,
this could even be used to cycle out offline peers for new online ones.

This process can take some time, since peers are only queried one by one
and may not even have a successful response. There is also a maximal
penalty applied because this happens at the very beginning of the
simulation, so reponses take a long time.

There is currently only a reputation system for query peers, but not for
sync peers. When a query is sent to a sync peer, it is treated as though
the querying peer has 0 reputation. This should be changed once the sync
group system is made into a proper part of the network that includes a
reputation system.

In the end, the changes seem to work. After a while, no more impossible
queries are observed where previously they didn't stop. Theoretically,
it should be possible to completely get rid off impossible queries as
long as the peer graph is strongly connected (in the current setup,
there is one component of 2 nodes that should be cut off from the rest,
but doesn't seem to be for some reasone though).