Consider own delay penalty
When responding to a query and calculating the delay with which to
respond, peers now consider the delay of the query they potentially had
to issue themselves so that the delay the querying peer perceives
ideally doesn't exceed the delay that was supposed to be imposed.

For example, say A issues a query to B and B decides that the delay
should be 10 time units. But B needs to issue a query to get the result
first, which takes 6 time units. B then only delays the response to A by
4 time units.

Of course, if the delay of the second-level query is larger than that of
the initial query, nothing can be done to go back in time. The
additionaly delay imposed then is 0.