Proxy action methods
The act_* members were previously partial function objects that
referenced act_*_default methods. The idea was that there would be other
versions for the implementation and a peer could periodically update the
partials in order to reevaluate its strategy.

But that doesn't work if the behavior depends on the parameters of the
method (e.g. whom a request is coming from). So instead of the partials,
there are now proper proxy methods in which decisions like that can be
made. Persistent strategy decisions will have to be explicitly stored in
variables somehow.

This commit does not change the behavior of peers.