Revise second presentation
Move lots of stuff to the backup slides, include concrete examples.
Includes log files of 2 runs of the prototype simulation with good vs.
selfish peers.

The animation illustrating query groups doesn't look great. The Ellipses
should probably have less opacity.

The histograms are also not great. For small bars, the actual number of
elements in the bin is printed above, but it doesn't look great.