Don't log query on request
When handling a request, recv_query() is called (with sender and
recipient being the peer handling the request). This had the side effect
of a request and an incoming query being logged at the same time.

Added a flag to the call that handle_request() can toggle to avoid
logging.