Link log events
All log events now have an in_event_id, which is the ID (the index in
the event list) of the event that caused the event. This way, all events
are linked in a causal chain.

There is as of yet no way to go forward in that chain, i.e. look at
events that were caused by an event, but that is possible with some
post-processing.

Implementing this required passing through the in_event_id in the peer
logic, which unfortunately made my try/finally solution to logging
unworkable. I've left this in (with the finally block being a pass) to
avoid a confusing diff.