Introduce addresses
Peers no longer have access to the reference of other peers. They must
register with the network to get an address, and to communicate with
another peer must resolve its ID to such an address. The
send_{query,response}() methods of the network abstraction take these
addresses as parameters.

Addresses are simply integers that are increasing as new ones are
registered. They are not reused, although this should be implemented in
the future so peers have to check that who they're talking to is really
who they think it is.

If a peer tries to send a message to an address that is not associated
to a peer, an exception is raised. This doesn't happen in the current
version though since peers never go offline.

Peers have a method to look up the address of a peer with a given ID
from ID/address mappings stored locally by the peer.  If a peer tries to
send a message to another peer where this resolution fails, currently a
NotImplementedError is raised. It would be possible to do another query
to look up the missing mapping before sending the message. This may
become necessary in the future, e.g. if ID/address mappings have a TTL
which has expired. But for now, the exception is never raised because
peers only send messages to other peers whose addresses they know.

There are 2 workaround hacks necessary for the moment:

1. When adding another peer to a query group (as the second or third
possibility in join_group_with()), the other peer's reference is needed
to update its set of query groups. This is done by accessing the network
abstraction's private mapping from peer IDs to references. This will
cease to be necessary once the query group invite system is implemented.

2. Similarly, to even find all the query groups another peer is in,
peers use the all_query_groups set. Again, the invite system should make
this unnecessary.

There are 2 new classes, PeerInfo and QueryPeerInfo, which are a handy
way of storing ID, prefix and address of peers.

Large parts of this commit are actually variable renamings e.g. from
peer to peer_id, for clarity.