Send reputation updates
Do reputation updates in-network. Upon receiving a response, or a
response timing out, a peer now sends messages to increase or decrease
the responding (or timed out) peer's reputation appropriately.

Peers do not lie about the amount of reputation credited or debited and
send these messages to all relevant peers.

Since 2 peers may share more than 1 query group, this update is sent to
all query groups containing the peer about whom it is. However, when
receiving such an update, peers should only modify the reputation in
groups containing both the sender and the subject of the update. If a
receiver of the update shares other query groups with the subject of the
update, he must not modify the reputation in those, since other peers in
them are likely not known by the sender of the update and will not
receive the update themselves. This would lead to desynchronization of
the reputation record.

At the moment, peers only modify reputation when they're sending and
update, not when receiving. This is a necessary workaround because the
query group objects are shared between peers and otherwise each update
would be applied multiple times.

There are now separate log events for sending and for receiving a
reputation update. There is another one just for the global information
(we can't use either the sending or receiving event because they happen
many times for each update). Reputation decay is still handled globally,
and there is another log event for it (since there is no sender or
receiver).

Currently, updates are applied in the order in which they arrive. This
is however not sufficient, as application order matters. E.g., say a
peer has 1 reputation and 2 updates arrive. The first subtracts 2, the
second adds 5. In that order, the peer ends up with 5 reputation. But if
the second update were to be applied first, it would have 4.

So when sending updates, peers have to include the current time. When
receiving one, they have to check that there was not an update that
arrived earlier that should be applied later. In that case, they need to
be able to roll back and apply the updates again in the correct order.
As a tiebreaker for equal times, reputation increases should be applied
first.