Apply reputation updates in-order
When sending reputation updates, peers now include the time at which the
update is meant to be applied (the current time). When receiving, peers
make sure that updates are applied in correct temporal order, rolling
back and reapplying already processed updates if necessary.

To do this, the query peer info now stores a log of recent reputation
updates that are used for rolling back. Old updates are pruned from this
log after a while. This must not be done too eagerly, because that could
result in reputation data becoming desynchronized between peers (if one
peer can roll back further than another).