Use deterministic dicts and sets
Iteration order over dicts and sets is not deterministic. This was
actually a problem, with different runs iterating over them in different
orders.

Instead of regular dicts, use collections.OrderedDict. This guarantees
the ordering of entries.

Instead of set, use util.SortedIterSet. This is just a subclass of set
that creates and sorts a list of its entries when returning an iterator.
This means a bit of overhead of course, but should be fine as long as
sets don't grow too big.

Sorting the set entries requires everything to going into sets to
implement __lt__(). This was done for bitstring.Bits (monkeypatched in
in the simulation module), Peer, QueryGroup and PeerInfo. The latter 3
use the peer's or query group's ID, the former sorts bitstrings first by
length, then by the unsigned integer value represented by them.