Fix unmatched queries
There were a couple of responses in the simulation results with status
unmatched, which shouldn't have been there. These occurred when peers
would send responses to themselves as a last step after they had
received a failed response before the query had been archived. This led
to there not being an archived query to match and no answer in the
response, leading to the unmatched status.

Peers now don't send responses to themselves at all. Instead, the new
QueryFinalized event marks the end of a query chain (i.e. a response
arriving at the peer where a request initially occurred).

IDs for this new type of event could also be stored in the archived
queries and later connected to possible late responses, but this isn't
done at the moment.