Introduce settings files
Instead of constants governing behavior, there is now a mechanism for
settings files. A file name must be passed as the first argument (with
the simulation duration an optional second).

The settings file is simply a string representation of a python
dictionary that is eval'd. An annotated example file with the default
settings is in the simulation directory.

Also fixed some of the unit tests.