Let peers change query groups
Peers periodically check how well query groups are performing for them.
If they don't they'll look for a replacement if necessary and then leave
it.

The decision whether a group has bad performance is based on how much
reputation the peer can get in it. The group is considered good if the
peer has a minimum reputation, or if he is gaining reputation quickly
enough. However, a peer will not leave a query group if he has just
joined it. In order to determine how quickly he is gaining reputation,
every peer regularly records his current reputation in every query
group. Then a linear regression is fitted through the most recent data
and the slope used to make the decision.

If a group is found not to perform, the peer will check what he needs it
for, i.e. if after removing it for any of his subprefixes he knows less
than the minimum desirable number of peers that serve the subprefix. If
he doesn't need it, he just leaves, if he does, he tries to find a
replacement first (another query group containing other peers covering
the subprefixes). If none can be found, he stays in the badly performing
group.

Finding a new group is done by accessing the shared all_query_groups
dict. This mechanism also makes find_missing_query_peers() useless for
the moment, since that actually sends out queries that take time, while
the new mechanism is instantaneous.

The decision which other group to choose is very primitive and really
not helpful. The peer simply takes the first query group from the dict
that is not full and has at least one peer he needs (and possibly more
if he needs more peers). And since the dict is actually and OrderedDict,
the first few groups get a lot of traffic. And since the peer doesn't
cosider at all how well the new group may perform, he might have to
leave the new group and then just goes back to the old one, ad inf.

Peers need to look at the group before joining and by some metric
determine that it will likely perform better than the one they're
leaving.

There is a new log event, QueryGroupReevaluation, that's logged
everytime this happens, its out_event_ids lead to the relevant
QueryGroup{Add,Remove} events.