Consider own usefulness before switching groups
Before joining a new query group as a replacement for an old one, peers
now check how useful they would be in it, as an estimate for how well
they would be gaining reputation there. If it is lower than in the
current group, they don't make the switch.

The usefulness is calculated as the number of other peers in the group
that need a subprefix served the peer serves.