Expect future penalties
Whenever a peer does something that will incur a penalty (a reduction in
reputation), he now remembers it for later. When deciding whether or not
to misbehave, he will consider all the penalties that are likely to
arrive in addition to his current reputation. This is supposed to
prevent peer's reputations from crashing when a lot of penalties arrive
at once.

The expected penalties are forgotten after a timeout.