Clear timed out penalties before matching
Peers now clear timed out entries from the list of expected penalties
before trying to match and incoming penalty to an entry in that list.

This is supposed to fix cases where peers would match the incoming
penalty to a record that has already timed out, and the correct match
timing out as well.