Don't expect penalties for unintentional behavior
Removed the expectation of penalties for behavior that is not
intentional. This happens in cases where a peer wants to respond in
time, but is unable because he needed to query some other peer to do it,
but that response is too late.

The behavior so far was to expect a penalty as the peer responds. This
mostly isn't helpful since by that time, the penalty has likely already
been applied (which will have been unexpected). The expected penalty
would then just time out.