Handle responses to multiple queries
In very rare circumstances, it can happen that a peer receives a
response that answers 2 queries at the same time that were stored as
separate pending queries. For this to happen, the peer must first send
out a query for a prefix (not a whole ID), and then a query for an ID
(or a longer prefix) that begins with this prefix. In this order, the
second query can not be subsumed by the first (because it's more
specific). Furthermore, these 2 queries must be sent to the same peer.
That peer may then send a single response in which the queried_ids set
contains the prefixes or IDs from both queries.

Previously, there was an assertion checking that only for one of the
elements in that set there exists a record in the pending queries
dictionary. This didn't present a problem until now because the
circumstances are very rare.

Updated recv_query() to be able to handle responses answering muliple
pending queries. In case of a retry, the pending queries are merged into
one.