Attenuate rewards
Added strategies to attenuate peers' rewards based on a peer's current
reputation. All of them specify a lower_bound and an upper_bound.
Reputation below the lower_bound is not modified, and reputation above
the upper_bound is not possible.

E.g., if with lower_bound == 10 and the current reputation == 9 a reward
of 5 arrives, 1 of it is applied without modification (to get to 10).
The remaining 4 are attenuated according to the strategy. This is always
done in a way such that it doesn't matter whether a reward of 4 is
applied at once, or 2 rewards of 2 are applied consecutively.
upper_bound simply acts as a reputation cap that cannot be surpassed.

The strategies are:

constant: All reputations above lower_bound are multiplied by a
coefficient.

square_root: The square root is taken of all reputation above
lower_bound and the result multiplied by a coefficient.

harmonic: This is like constant, with the coefficient getting stepwise
smaller as the current reputation increases. The parameters a and k
govern how fast this happens. Reputation between lower_bound + i and
lower_bound + i + 1 is multiplied by 1 / (a + i * k). So with k == 0,
this is the same as constant, where a is the inverse of the coefficient.

These strategies only apply to rewards (i.e. positive reputation
changess), penalties are not touched, nor is the decay rate.