Fix rare retry issue
This bug never exhibited until now because it's just so rare. It only
happens when a query for a subprefix is pending, then a request comes in
for a longer but matching subprefix that's sent to the same peer and
then the response is unsuccessful but can be retried. Or something like
that.