Ignore non-existent subprefixes
Peers have the option to ignore subprefixes for which there doesn't
exist any peer in the simulation when querying for missing subprefixes.
It requires some global knowledge given to them by the simulation that
wouldn't be available to a peer in a real system.

The reason for this is to fix most peers' reputations crashing when the
prefix length is so long that some subprefixes just don't have any
peers. What would happen is that peers send out queries for them that
immediately fail and are penalized. This goes on and keeps the
reputation of those peers receiving these queries low.

The underlying problem is that the system doesn't yet have a sensible
way of dealing with queries for non-existent IDs.