Don't query sync peers by default
When finding peers for their subprefixes, peers used to query all peers
they knew, including peers who are further away from the subprefix,and
even sync peers (there was no way to do it properly and query peers that
are closer to the ID they need, since they don't have coverage for that
subprefix).

This patch allows to make a distionction between peers who are further
away and sync peers. Sync peers are not queried by default now.

This addresses the problem that, with the current design, sync peers
don't actually have any incentive to respond. It should work with only
query peers.