Guarantee sync groups not being empty
When creating peers for the simulation, there is now an option to ensure
that sync groups are non-empty, i.e. that there exists at least one peer
for every possible prefix.

The peers with different sync groups are created in the beginning so
that, even if they are created in multiple batches, there are no empty
sync groups as soon as possible. However, if the first batch or batches
are too small, there will still be some empty sync groups for a while.