Overhaul query management
Changed the way peers store information about queries they receive and
send. Previously, when a peer received a query, he would create a
PendingQuery. This held information about what was queried for, who
queried for it (peers could be added if more queries came in), what
other peers should be and have been queried. That became a bit unwieldy.

Now peers maintain separate maps for incoming and outgoing queries
(stored as a trie using pytrie). When a query comes in, the peer records
this and sends out a query unless one is already underway. When a
response arrives, the peer gets the record of the outgoing query and
either deletes it (on success) and answers all matching incoming
queries, or (on failure) updates it so that the peer that failed is not
queried again and retries.

This system is much more flexible. Another major advantage is that the
list of peers that should be queried to find a record doesn't need to be
created at the very beginning and can't change over time. The outgoing
query just notes what peers have already been queried, and they are not
tried again. So peers use up-to-date information when they send a query,
as query groups change.

However, there is now no easy way to distinguish between a query failing
ultimately, and a query failing that can be retried. But that wasn't
very useful anyway.

It will be easier to make peers send out multiple queries.

Peers no longer store outgoing queries when they have timed out, so they
no longer give rewards for queries that are correct but late. This may
or may not be a good thing.

Peers now respond to exactly one queried ID, even if that means sending
out 2 responses with the same information (one of the queried IDs is a
prefix of the other). This was just easier, is a rare case and doesn't
detract from the simulation.

Slightly unrelatedly, there is now a distinction in a peer's response
between the peer saying he doesn't know the answer to the query, and
saying there is no matching peer. This already successfully removes the
problem of massive penalties incurred by correctly claiming there are no
peers with a prefix during finding missing query peers in the beginning.

But now peers also need to decide whether they trust a peer that claims
a queried-for peer doesn't exist, or if they want to confirm with a
second query to someone else. This has not been implemented.