Specify known peers on query
When querying for a prefix, peers now have the option to specify a set
of peer IDs that they already know and that should not be returned. This
makes it possible to find out whether there even exists a peer covering
a subprefix: if there isn't, the responding peer will answer that no
such peer exists.

There is one major problem with this still unaddressed: If the prefix
that is queried for is shorter than the prefix length, there are
multiple sync groups from which every single peer is a valid answer. If
a peer receives a query for it and all his sync peers are excluded, he
can't return any of them. But he also can't say that there doesn't exist
a peer with the queried prefix since he doesn't know the other sync
groups. If he queries another peer from one of the other sync groups,
that peer will have the same problem of not knowing any of the other
sync groups. The solution would be to exclude the entire sync group via
prefix, so that the peer from another sync group knows not to send the
query back and is able to respond that there doesn't exist a peer with
the queried prefix.