Exclude prefixes
Peers can now exclude prefixes. This signals to the peer they query that
their response's ID should not start with that prefix.

This makes it possible for peers to exclude their entire prefix (their
sync group) to signal that there is no peer in their sync group that can
satisfy the query, thus solving the earlier problem of possible cyclical
queries when finding missing subprefix coverage.

A way for peers to actually ultimately signal that there is no peer
satisfying the request is still to be implemented though.

The set of excluded peer IDs is now represented as a trie (with all
values None) to enable prefix searches.