Fix double query issue
There was a rare issue where peers would send a query to another peer
for an ID for which there already was a query underway. This led to
send_query() overwriting the OutgoingQuery, losing the timeout process
which then couldn't be interrupted anymore. When one of the queries
finished, it would be deleted from the outgoing query storage. When then
the other query timed out, no record could be found of it.

This was introduced with the ability to exclude prefixes. Previously,
there was a check that there wasn't already a query for the exact ID to
the same recipient underway, but that had been modified to also account
for matching excluded peer IDs, meaning that, in some cases where they
didn't match, it was possible for basically the same query to be sent
twice.

The clean solution would be to enable peers to track multiple queries
for the same ID to the same peer that only differ in the set of excluded
peer IDs.

However, this is a disproportionate amount of work considering how rare
this issue is likely to be, so this fix simply assures that a peer to
whom a query is in progress is not queried for the same ID until he
answers, even if the excluded peer IDs don't match. This is done by
adding all such peers to the set of peers already queried.