Work around double incoming queries
This is a fix analogous to the earlier commit that dealt with multiple
outgoing queries to the same peer for the same ID where only the set of
excluded peer IDs differed. This one is basically the same for incoming
queries.

In case a query arrives from a peer for an ID for which an incoming
query record already exists, but the query has to be done anyway because
the new query has more restrictive excluded peer IDs, no new
IncomingQuery is created anymore (which would overwrite the old one).
Instead, the old one is fetched and the excluded peer IDs updated to
also satisfy the new incoming query (so the one existing object now
satisfies both of the queries the peer received). The further query that
is still necessary is done as usual.

The minor downside of this is that the second one of the queries is
answered with a delay that is appropriate for the first of them. But
without a fairly comprehensive overhaul, again in relation to how rare
this issue is likely to be, peers have no means of storing that
information.