Ignore responses not matching exclusions
Sometimes a peer may receive a response that doesn't honor the excluded
peer ID requirement of the outgoing query on record. The circumstances
for this are thus:

Peer A sends a query to peer B, which times out, thus deleting the
outgoing query record. A then sends a new query for the same ID, again
to B (this can happen in find_missing_query_peers(), it's a completely
new request and so B isn't in peers_already_queried), but with a more
restrictive excluded peer set. Before this new query arrives, B responds
to the first query with an answer that's fine for the first query, but
doesn't honor the more restrictive peer exclusion of the second. A then
receives this and, having lost record of the first query, thinks it's
supposed to be an answer to the second one.

Previously, it was asserted that this can't happen (indeed, it couldn't,
because this is again a problem of peers being unable to track outgoing
queries that differ only in the restrictiveness of excluded peers). Now,
it's simply logged as an inappropriate response. The correct query will
still be underway, and the actual response should still arrive.