Respond "no such peer"
Peers now finally have the ability to claim in a response that a peer
that was queried for (or any peer with a prefix that was queried for)
doesn't exist.

There are 2 cases where this can happen:

1. The queried ID is at least as long as the prefix length (e.g. a full
   ID) and should be in the peer's sync group. If the peer doesn't see
   any matching peer in the sync group, he can claim there doesn't exist
   such a peer.

2. The queried ID is shorter than the prefix length, there is no match
   in the peer's sync group, and all other prefixes (covering all other
   sync groups) have already been excluded. In this case, the peer knows
   there is no match in his sync group, and no match from any other sync
   group would be accepted.

This makes it possible for peers to find out that e.g. there only exists
one peer for a particular subprefix they need covered, even though they
might want more.

Cases in which this fails is if there doesn't exist any peer for some
sync group. Then there is no one in that group that can claim that there
is no one in it (only sync group members are authorities on who is in a
sync group). This is a problem related to a fragmented network: maybe
there are peers in that sync group, but none of them knows anyone in my
part of the network.