Write plots to file directly
Logger.save_plots_until() calls all plotting functions of Logger and
writes them to files directly for convenience. Takes the simulation time
until which the log should be processed and a file name prefix.
Appropriate suffixes for the plot types are appended.

The image size may be too small, we'll see.