Use proper half columnwidth
I used to use 0.49\columnwidth as the width for two graphics side by
side, because with 0.5 they were displayed one above the other. That was
because of a spurious space at the end of the first, which made the line
longer. That space can be suppressed with a comment at the end of the
first graphic.